requirements.txt
Django==4.1.10
djangorestframework==3.14.0
django-cors-headers==4.0.0
celery==5.3.4
redis==5.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
